added new routers and moved server.js into server folder
