+++ server/old_helpers/qanda-helpers.js
